Make epoll runtime configurable and disable by default
Storage server is spitting out some occassional errors/crashes with this
enabled, so this makes it opt-in rather than always on.

This effectively makes it experimental, which is fine (the only place
this has any advantage is in the SPNS, and that will likely transition
to quic in the not-too-distant future).