Ensure service node status is set on incoming commands
If an incoming connection gets established before the local node can
detect it as a service node (for instance: because it is still syncing
the chain and is behind) then that peer won't have `service_node` set,
and it might remain unset if the peer remains connected.

This fixes it to allow upgrading the peer to set the service node flag,
which ought to eliminate the (wrong) `remote is not recognized as a
service node` log warnings.