Fix destruction hang after start() failure with tagged threads
If startup fails (e.g. because a bind failed in the proxy thread
initialization) and there are worker threads then the OxenMQ destructor
would hang waiting because the worker thread destructors would never
finish.

This hang would show up in oxend if oxenmq binding failed.

This fixes it by properly re-joining the proxy thread during start() if
it throws before we propagate the exception so that we properly enter
the !joinable part of the destructor to properly clean up the tagged
workers.