Fix inproc connections on active SNs
When the current node is in the active service node list, an inproc
connection wasn't populating `peers` list, but when attempting to reply
to the message, since it was a SN, we expected to find it in `peers`.

As a consequence, oxend interactive mode commands were not seeing the
inproc connection when attempting to route the reply, and so attempted
to do a remote address lookup and connection, which meant 1) interactive
commands pointlessly opened a new external connection, and 2) if the
connection failed (or there was no IP) then the command would fail
completely and all interactive commands would just time out with errors
about the pubkey being unreachable.

This fixes it by making inproc connections populate the peers list in
the same way as other types of incoming connection so that we find and
use that connection when replying to the local rpc commands.