Migrate oxenmq logging to oxen-logging
Replace oxenmq's build in callback logger and log level with
oxen-logging log messages.

This makes it much easier to integrate with everything else in the stack
which are now all using oxen-logging.

Unfortunately this breaks the API because all the existing log callbacks
being passed to constructors and whatnot break, so this bumps the
version to 1.3.0.

Also bumps Catch2 to the latest stable v3 version, updates the tests to
use it, and to support --log-level and --log-file to configure oxenmq
logging in the tests.